Унифицировал получение scriptVariant и compatibilityMode
